requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Gemini (Google AI)
langchain-google-genai==2.0.5
langchain-core==0.3.28
httpx>=0.27.0

# LangChain RAG
langchain==0.3.7
langchain-huggingface==0.1.2
langchain-chroma==0.1.4
langchain-text-splitters==0.3.2
chromadb==0.5.23
sentence-transformers==3.3.1

# PDF
pdfplumber==0.11.0

# Utility
pydantic==2.7.4
python-dotenv==1.0.0